some cleanup, added some comments, moved some types around
